+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -21,6 +21,7 @@ public class ExperimentalRuleSetProvider : RuleSetProvider {
         SpacingAroundAngleBracketsRule(),
         SpacingAroundUnaryOperatorRule(),
         AnnotationSpacingRule(),
-        UnnecessaryParenthesesBeforeTrailingLambdaRule()
+        UnnecessaryParenthesesBeforeTrailingLambdaRule(),
+        BlockCommentInitialStarAlignmentRule()
     )
 }